Import xrandr.c as an implementation example.
Over the course of this development effort, code will be refactored out
of xrandr.c into the library source.  As this is done, the corresponding
code will be excised from this copy of xrandr.c and replaced with the
new API calls.  This commit pulls in the unchanged xrandr.c from the
xrandr package.

This will enable before/after testing of xrandr behavior to catch
functional regressions.

Ultimately, prior to libXrandrUtils 1.0.0, this work should be complete
and xrandr.c migrated back home to the xrandr package.

Signed-off-by: Bryce Harrington <bryce@canonical.com>